# packages/nexus-agent-core/capabilities/resource/openapi/control-plane/virtual-keys.yaml
...
openapi: 3.1.0
info:
    title: Nexus Control Plane Admin API — virtual-keys
    version: dev
    description: 'Hand-maintained against control-plane source: field names, types, optionality, status codes and IAM tiers mirror the handler code; enum value sets and required-field rules enforced imperatively in handlers are verified by the openapi-review skill.'
paths:
...
    /api/admin/virtual-keys/{id}/approve:
        post:
            operationId: approveVirtualKey
            summary: Approve a pending virtual key
            description: |-
                Approves a virtual key that is in "pending" status, transitioning it to
                "active" and recording the approver. Returns 404 when the key does not
                exist or is not pending.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbApprove)
            x-nexus-tier: confirm
            parameters:
                - name: id
                  in: path
                  description: Virtual key id.
                  required: true
                  schema:
                    type: string
            responses:
                "404":
                    description: Virtual key not found or not in pending status.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    message:
                                        type: string
                            example:
                                message: Virtual key approved
...
    /api/admin/virtual-keys/{id}/reject:
        post:
            operationId: rejectVirtualKey
            summary: Reject a pending virtual key
            description: |-
                Rejects a virtual key that is in "pending" status, transitioning it to
                "rejected" and recording the rejecter and the supplied reason. The reason
                is stored as-is and is not validated as non-empty. Returns 404 when the key
                does not exist or is not pending.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbReject)
            x-nexus-tier: confirm
            parameters:
                - name: id
                  in: path
                  description: Virtual key id.
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                reason:
                                    type: string
                                    description: Free-text rejection reason recorded on the key.
                        example:
                            reason: Project scope not approved by security review.
            responses:
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "404":
                    description: Virtual key not found or not in pending status.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    message:
                                        type: string
                            example:
                                message: Virtual key rejected
    /api/admin/virtual-keys/{id}/renew:
        post:
            operationId: renewVirtualKey
            summary: Renew an active virtual key
            description: |-
                Extends the expiry of an active application virtual key. `expiresAt` is
                required and must be in the future; the distance is not bounded.
                The data plane cache is invalidated so the new expiry takes effect on the
                next request. Returns 404 when the key is not found, not active, or not an
                active application key.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbRenew)
            x-nexus-tier: confirm
            parameters:
                - name: id
                  in: path
                  description: Virtual key id.
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                expiresAt:
                                    type: string
                                    format: date-time
                                    description: New expiry. Must be in the future; the distance is not bounded.
                            required:
                                - expiresAt
                        example:
                            expiresAt: 2026-04-15T00:00:00Z
            responses:
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "404":
                    description: Virtual key not found, not active, or not an active application key.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    message:
                                        type: string
                                    expiresAt:
                                        type: string
                                        format: date-time
                            example:
                                message: Virtual key renewed
                                expiresAt: 2026-04-15T00:00:00Z
    /api/admin/virtual-keys/{id}/revoke:
        post:
            operationId: revokeVirtualKey
            summary: Revoke an active virtual key
            description: |-
                Revokes an active virtual key, transitioning it to "revoked" and
                invalidating it on the data plane. Returns 404 when the key does not exist
                or is not in active status.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbRevoke)
            x-nexus-tier: confirm
            parameters:
                - name: id
                  in: path
                  description: Virtual key id.
                  required: true
                  schema:
                    type: string
            responses:
                "404":
                    description: Virtual key not found or not in active status.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    message:
                                        type: string
                            example:
                                message: Virtual key revoked
...